requirement.txt
pymodbus==3.6.9
fastapi==0.110.0
uvicorn[standard]==0.30.1
pymongo==4.7.3
websockets==12.0
python-dotenv==1.0.1
pyserial==3.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openpyxl==3.1.2

